pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skyline-prism"
version = "0.1.0"
description = "PRISM (Proteomics Reference-Integrated Signal Modeling) - RT-aware normalization for Skyline proteomics data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "MacCoss Lab", email = "maccoss@uw.edu" },
]
keywords = [
    "proteomics",
    "mass-spectrometry",
    "normalization",
    "quantification",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.21",
    "pandas>=1.4",
    "scipy>=1.7",
    "scikit-learn>=1.0",
    "pyarrow>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
viz = [
    "matplotlib>=3.5",
    "seaborn>=0.12",
    "plotnine>=0.12",
]

[project.scripts]
prism = "skyline_prism.cli:main"

[project.urls]
Homepage = "https://github.com/maccoss/skyline-prism"
Documentation = "https://github.com/maccoss/skyline-prism#readme"
Repository = "https://github.com/maccoss/skyline-prism"
Skyline = "https://skyline.ms"

[tool.hatch.build.targets.wheel]
packages = ["skyline_prism"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=skyline_prism --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
